Add files needed to run the (?)breadboard
This is something the EE side of our team was involved in. They made a breadboard with three lights on it - red, blue, and green. They wrote up the script to light up red if the disposable was trash, blue if it was recyclable, and green if it were compostable.